git-svn-id: https://www.soclib.fr/svn/trunk/soclib@2261 2cf4c1f0-a32b-0410-9068-d80f8be93c97

+++ soclib/communication/vci/caba/source/include/vci_signals.h
@@ -143,10 +143,10 @@ public:
                       << "/wdata = " << wdata
                       << "/be = " << be
                       << std::dec
-                      << "/srcid = " << (uint32_t)srcid
-                      << "/trdid = " << (uint32_t)trdid
-                      << "/pktid = " << (uint32_t)pktid
-                      << "/plen = " << (uint32_t)plen 
+                      << "/srcid = " << srcid
+                      << "/trdid = " << trdid
+                      << "/pktid = " << pktid
+                      << "/plen = " << plen 
                       << "/eop = "   << eop
                       << "/ack = "   << cmdack << std::endl;
         }
@@ -156,9 +156,9 @@ public:
             std::cout << " rerror = "   << rerror
                       << "/rdata = "  << rdata
                       << std::dec
-                      << "/rsrcid = " << (uint32_t)rsrcid
-                      << "/rtrdid = " << (uint32_t)rtrdid
-                      << "/rpktid = " << (uint32_t)rpktid
+                      << "/rsrcid = " << rsrcid
+                      << "/rtrdid = " << rtrdid
+                      << "/rpktid = " << rpktid
                       << "/reop = "   << reop
                       << "/ack = "    << rspack << std::endl;
         }
